Add support for SSL/TLS for Elasticsearch HTTP and Transport layers (#187)
Add support for SSL/TLS for Elasticsearch HTTP and Transport layers

This commit adds support for securing HTTP and Transport layers by specifying parameters with which PKCS#12 archives containing the certificate and key can be supplied. 

To configure TLS for the HTTP layer, a PKCS#12 archive containing the HTTP cert and key or an archive containing the HTTP CA cert and key must be provided. 

When an archive containing the HTTP cert and key is provided, it will take preference, and Kibana will be configured with certificate verification mode. The most likely use case for this is to provide a cert for a CNAME pointing at the external loadbalancer public IP.

When an archive containing the HTTP CA cert and key HTTP certs is provided, the CA certs is used to generate a PKCS#12 archive containing a certificate and key for each node, including the node DNS and IP as Subject Alternative Names, as well as the internal loadbalancer public IP. This allows Kibana to be set to full verification mode when communicating internally. A client
communicating through the external loadbalancer can verify certificates returned from Elasticsearch against the CA. A client communicating through Application Gateway will use the certificate configured for the Application Gateway.

To configure TLS for the Transport layer, a PKCS#12 archive containing the Transport CA cert and key must be provided. The CA cert is used to generate a PKCS#12 archive containing a certificate and key for each node, including the node DNS and IP as Subject Alternative Names. Including the IP allows the cluster to operate in full verification mode for the Transport layer. 

Other changes:

- Use the insecure flag when calling localhost over HTTPS with curl. Since the subject name in certificate used to secure the HTTP layer is most likely not going to match localhost (it's likely to be tied to a public domain name), the --cacert flag for curl cannot be used.

- Update integration tests to use the parameters-file argument to pass the parameters to azure cli.
    Since parameters can now contain base 64 encoded certificates, the input can be longer than the maximum characters allowed in Windows (8192).

- Add generated self-signed certificates for use in integration tests.
- Add --test parameter to be able to filter integration tests
    --test <regexp> with a pattern to match the tests to run
- Configure Application Gateway to work with TLS on backend pool
    Add support for TLS from Application Gateway to the backend pool. Application Gateway communicates with the backend pool through the internal loadbalancer, and the public certificate(s) used by the backend pool must be whitelisted by Application Gateway by providing these details to it.
    This means that for TLS on the HTTP layer in conjunction with Application Gateway,
only the single esHttpCertBlob option can be supported. In the case of esHttpCaCertBlob which is used to generate a cert for HTTP layer for each VM/node in the cluster, the public keys for these certs cannot be automatically added to Application Gateway as part of deployment.

- Pass Elasticsearch HTTP and HTTP CA archives for Kibana configuration
    Pass the Elasticsearch HTTP and HTTP CA PKCS#12 archives to Kibana to configure TLS to Elasticsearch from Kibana. When a CA cert is provided, it is used to configure the certificate authority.
    The presence of an Elasticsearch HTTP archive overrides the presence of a HTTP CA archive. That is, if a HTTP archive is provided, the Certificate Authority will be extracted from this PKCS#12 archive.
If no CA cert is present, The verification mode for TLS with Elasticsearch will be set to none.

- Remove versions less than 5.3.x from the template.
    5.0.x and 5.1.x are now EOL and 5.2.x will be EOL end of July.
    5.2.x is removed now because Kibana Console does not work with self-signed
certs; Kibana itself works, but not console, responding with

    Client request error: unable to verify the first certificate

    Closes #199

- Remove hostname verification in integration tests

    Remove the hostname verification check used by node's request module. Since all certs used are self-signed, hostname verification will fail. Tests still verify the CA.

- Add support for --nodestroy parameter

    Adds a --nodestroy parameter that when passed, does not delete resource groups after integration tests

+++ build/tasks/arm-validator.js
@@ -2,6 +2,7 @@ var config = require('../.test.json');
 var gulp = require("gulp");
 const execFile = require('child_process').execFile;
 var fs = require('fs');
+var path = require('path');
 var _ = require('lodash');
 var dateFormat = require("dateformat");
 var git = require('git-rev')
@@ -12,10 +13,9 @@ var request = require('request');
 var hostname = require("os").hostname();
 var operatingSystem = require("os").platform();
 var compareVersions = require('compare-versions');
+var argv = require('yargs').argv;
 
-var azureCli = "../node_modules/.bin/azure";
-if (operatingSystem === "win32")
-  azureCli = "..\\node_modules\\.bin\\azure.cmd";
+var azureCli = (operatingSystem === "win32")? "..\\node_modules\\.bin\\azure.cmd" : "../node_modules/.bin/azure";
 var artifactsBaseUrl = "";
 var templateUri = "";
 var armTests = {};
@@ -42,6 +42,22 @@ var exampleParameters = require("../../parameters/password.parameters.json");
 var bootstrapTest = (t, defaultVersion) =>
 {
   var test = require("../arm-tests/" + t);
+
+  // replace cert parameters with values with base64 encoded certs
+  [
+    "esHttpCertBlob",
+    "esHttpCaCertBlob",
+    "esTransportCaCertBlob",
+    "kibanaCertBlob",
+    "kibanaKeyBlob",
+    "appGatewayCertBlob",
+    "appGatewayEsHttpCertBlob"].forEach(k => {
+    if (test.parameters[k] && test.parameters[k].value) {
+      var cert = fs.readFileSync("certs/" + test.parameters[k].value);
+      test.parameters[k].value = new Buffer(cert).toString("base64");
+    }
+  });
+
   log(t, `parameters: ${JSON.stringify(test.parameters, null, 2)}`);
   var testParameters = merge.recursive(true, exampleParameters, test.parameters);
   testParameters.artifactsBaseUrl.value = artifactsBaseUrl;
@@ -55,24 +71,43 @@ var bootstrapTest = (t, defaultVersion) =>
   testParameters.securityLogstashPassword.value = config.deployments.securityPassword;
   testParameters.esVersion.value = defaultVersion;
 
+  // Some parameters are longer than the max allowed characters for cmd on Windows.
+  // Persist to file and pass the file path for parameters
+  var resourceGroup = "test-" + hostname.toLowerCase() + "-" + t.replace(".json", "") + dateFormat(new Date(), "-yyyymmdd-HHMMssl").replace("+","-")
+  var testParametersFile = path.resolve(logDistTmp + "/" + resourceGroup + ".json");
+  fs.writeFileSync(testParametersFile, JSON.stringify(testParameters, null, 2));
+
   return {
-    resourceGroup: "test-" + hostname.toLowerCase() + "-" + t.replace(".json", "") + dateFormat(new Date(), "-yyyymmdd-HHMMssl").replace("+","-"),
+    resourceGroup: resourceGroup,
     location: test.location,
     isValid: test.isValid,
     why: test.why,
     deploy: test.deploy,
-    params: testParameters
+    params: testParameters,
+    paramsFile: testParametersFile
   }
 }
 
 var bootstrap = (cb) => {
   var allowedValues = require('../allowedValues.json');
-  var defaultVersion = _.last(allowedValues.versions);
+  var defaultVersion = argv.version ?
+    argv.version == "random" ?
+      _.sample(allowedValues.versions)
+      : argv.version
+    : _.last(allowedValues.versions);
+
+  if (!_.includes(allowedValues.versions, defaultVersion)){
+    return bailOut(new Error("No version in allowedValues.versions matching " + defaultVersion));
+  }
+
+  var templateMatcher = new RegExp(argv.test || ".*");
+
   git.branch(function (branch) {
     artifactsBaseUrl = "https://raw.githubusercontent.com/elastic/azure-marketplace/"+ branch + "/src";
     templateUri = artifactsBaseUrl + "/mainTemplate.json";
     log(`Using template: ${templateUri}`, false);
     armTests = _(fs.readdirSync("arm-tests"))
+      .filter(t => templateMatcher.test(t))
       .indexBy((f) => f)
       .mapValues(t => bootstrapTest(t, defaultVersion))
       .value();
@@ -150,8 +185,19 @@ var deleteAllTestGroups = function (cb) {
 var deleteCurrentTestGroups = function(cb)
 {
   var groups = _.valuesIn(armTests).map(a=>a.resourceGroup);
-  log(`deleting current run groups: ${groups}`, true);
-  deleteGroups(groups, cb);
+  if (argv.nodestroy) {
+    log(`not destroying ${groups.length} resource groups as --nodestroy parameter passed.`, false);
+    log("----------------------------------------------");
+    log("| DELETE THESE RESOURCE GROUPS WHEN FINISHED |");
+    log("|                                            |")
+    log("| $ npm run azure-cleanup                    |")
+    log("----------------------------------------------");
+    cb();
+  }
+  else {
+    log(`deleting current run groups: ${groups}`, true);
+    deleteGroups(groups, cb);
+  }
 }
 
 var createResourceGroup = (test, cb) => {
@@ -186,13 +232,12 @@ var validateTemplates = function(cb) {
 
 var validateTemplate = (test, cb) => {
   var t = armTests[test];
-  var p = JSON.stringify(t.params);
   var rg = t.resourceGroup;
   createResourceGroup(test, () => {
     var validateGroup = [ 'group', 'template', 'validate',
       '--resource-group', rg,
       '--template-uri', templateUri,
-      '--parameters', p,
+      '--parameters-file', t.paramsFile,
       '--json'
     ];
     log(`validating ${test} in resource group: ${rg}`);
@@ -253,10 +298,10 @@ var sanityCheckDeployment = (test, stdout, cb) => {
 
   if (stdout) {
     var outputs = JSON.parse(stdout).properties.outputs;
-    if (outputs.loadbalancer.value !== "N/A")
+  if (outputs.loadbalancer.value !== "N/A")
       checks.push(()=> sanityCheckExternalLoadBalancer(test, "external loadbalancer", outputs.loadbalancer.value, allChecked));
-    if (outputs.kibana.value !== "N/A")
-      checks.push(()=> sanityCheckKibana(test, outputs.kibana.value, allChecked));
+  if (outputs.kibana.value !== "N/A")
+    checks.push(()=> sanityCheckKibana(test, outputs.kibana.value, allChecked));
   }
 
   if (t.params.loadBalancerType.value === "gateway")
@@ -298,8 +343,31 @@ var sanityCheckExternalLoadBalancer = (test, loadbalancerType, url, cb) => {
   var t = armTests[test];
   var rg = t.resourceGroup;
   log(`checking ${loadbalancerType} ${url} in resource group: ${rg}`);
-  var superuser = compareVersions(t.params.esVersion.value,'5.0.0') >= 0 ? "elastic" : "es_admin";
-  var opts = { json: true, auth: { username: superuser, password: config.deployments.securityPassword }, strictSSL: false };
+  var opts = {
+    json: true,
+    auth: { username: "elastic", password: config.deployments.securityPassword },
+    agentOptions: { checkServerIdentity: _.noop }
+  };
+
+  var certParams = {
+    blob: (loadbalancerType === "application gateway") ? "appGatewayCertBlob": "esHttpCertBlob",
+    passphrase: (loadbalancerType === "application gateway") ? "appGatewayCertPassword": "esHttpCertPassword",
+  };
+
+  if (t.params[certParams.blob] && t.params[certParams.blob].value) {
+    if (t.params[certParams.passphrase] && t.params[certParams.passphrase].value) {
+      opts = merge.recursive(true, opts, {
+        pfx: fs.readFileSync("certs/cert-with-password.pfx"),
+        passphrase: t.params[certParams.passphrase].value,
+      });
+    }
+    else {
+      opts = merge.recursive(true, opts, {
+        pfx: fs.readFileSync("certs/cert-no-password.pfx")
+      });
+    }
+  }
+
   request(url, opts, (error, response, body) => {
     if (!error && response.statusCode == 200) {
       log(test, `loadBalancerResponse: ${JSON.stringify(body, null, 2)}`);
@@ -338,8 +406,18 @@ var sanityCheckKibana = (test, url, cb) => {
   var t = armTests[test];
   var rg = t.resourceGroup;
   log(`checking kibana at ${url} in resource group: ${rg}`);
-  var superuser = compareVersions(t.params.esVersion.value,'5.0.0') >= 0 ? "elastic" : "es_admin";
-  var opts = { json: true, auth: { username: superuser, password: config.deployments.securityPassword } };
+  var opts = {
+    json: true,
+    auth: { username: "elastic", password: config.deployments.securityPassword },
+    agentOptions: { checkServerIdentity: _.noop }
+  };
+
+  if (t.params.kibanaCertBlob && t.params.kibanaCertBlob.value) {
+    opts.ca = (t.params.kibanaKeyPassphrase && t.params.kibanaKeyPassphrase.value) ?
+      fs.readFileSync("certs/cert-with-password-ca.crt") :
+      fs.readFileSync("certs/cert-no-password-ca.crt");
+  }
+
   request(`${url}/api/status`, opts, function (error, response, body) {
     var state = (body)
       ? body.status && body.status.overall
@@ -360,12 +438,11 @@ var sanityCheckKibana = (test, url, cb) => {
 var deployTemplate = (test, cb) => {
   var t = armTests[test];
   if (!t.isValid || !t.deploy) return;
-  var p = JSON.stringify(t.params)
   var rg = t.resourceGroup;
   var deployGroup = [ 'group', 'deployment', 'create',
     '--resource-group', rg,
     '--template-uri', templateUri,
-    '--parameters', p,
+    '--parameters-file', t.paramsFile,
     '--quiet',
     '--json'
   ];
@@ -389,7 +466,6 @@ gulp.task("test", ["clean"], function(cb) {
 
 gulp.task("deploy-all", ["clean"], function(cb) {
   login(() => validateTemplates(() => deployTemplates(() => deleteCurrentTestGroups(() => logout(cb)))));
-  //login(() => validateTemplates(() => deployTemplates(() => () => logout(cb))));
 });
 
 gulp.task("azure-cleanup", ["clean"], function(cb) {
